fix: address PR task review feedback
- wait for terminal task state before returning tasks/result payloads
- reject cancel requests for inactive terminal jobs
- avoid task cursor loops and clean up progress notifier lifecycle
- move task tests to sidecar modules and split server metadata handlers
- refresh MCP task docs and remove unrelated session artifact

Resolves review thread PRRT_kwDORS2O8s6G-qWW
Resolves review thread PRRT_kwDORS2O8s6G-qWb
Resolves review thread PRRT_kwDORS2O8s6G-rzE
Resolves review thread PRRT_kwDORS2O8s6G-rzh
Resolves review thread PRRT_kwDORS2O8s6G-rzs
Resolves review thread PRRT_kwDORS2O8s6G-rz_
Resolves review thread PRRT_kwDORS2O8s6G-thT
Resolves review thread PRRT_kwDORS2O8s6G-thZ
Resolves review thread PRRT_kwDORS2O8s6G-thb
Resolves review thread PRRT_kwDORS2O8s6G-thf
Resolves review thread PRRT_kwDORS2O8s6G-thl
Resolves review thread PRRT_kwDORS2O8s6G-tho
Resolves review thread PRRT_kwDORS2O8s6G-thu
Resolves review thread PRRT_kwDORS2O8s6G-th2
Resolves review thread PRRT_kwDORS2O8s6G-th5
Resolves review thread PRRT_kwDORS2O8s6G-tiE
